Writing test in process 2
