+++ examples/mysql/20210315171956-create-users-table-up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS users
 (
-    id bigserial not null primary key,
+    id integer not null primary key auto_increment,
     name varchar (255) not null,
     email  varchar (255) not null unique,
     created_at timestamp DEFAULT CURRENT_TIMESTAMP
